server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Matěj Ščerba <m.scerba@seznam.cz>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
pandas = {version = "*", source = "pypi"}
flask = {version = "*", source = "pypi"}
structlog = {version = "*", source = "pypi"}
flask-cors = {version = "*", source = "pypi"}
pandas-stubs = {version = "*", source = "pypi"}
types-flask-cors = {version = "*", source = "pypi"}
pydantic = {version = "*", source = "pypi"}
flask-session = {version = "*", source = "pypi"}
redis = {version = "*", source = "pypi"}
psycopg2-binary = {version = "*", source = "pypi"}

[tool.poetry.dev-dependencies]
flake8 = {version = "*", source = "pypi"}
flake8-pyproject = {version = "*", source = "pypi"}
black = {version = "*", source = "pypi"}
mypy = {version = "*", source = "pypi"}
pre-commit = {version = "*", source = "pypi"}
types-psycopg2 = {version = "*", source = "pypi"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
ignore = ["E203", "W503"]
